requirements.txt
azure-ai-inference==1.0.0b9
azure-core==1.35.1
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
isodate==0.7.2
Markdown==3.10
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
